making the getBalance method return a double

+++ Labs/BankClass.cpp
@@ -28,7 +28,7 @@ class BankAccount
                 balance -= amount;
             }
         }
-        int getBalance()
+        double getBalance()
         {
             return balance;
         }
